moves common features and preset to different files
